Adding Etag

+++ src/MongoDbStorage.ts
@@ -1,5 +1,5 @@
 import { Storage, StoreItems } from 'botbuilder';
-import { MongoClient, Collection } from 'mongodb';
+import { MongoClient, Collection, ObjectID } from 'mongodb';
 
 export interface MongoDbStorageSettings {
   url: string;
@@ -58,28 +58,44 @@ export class MongoDbStorage implements Storage {
     const operations = [];
 
     Object.keys(changes).forEach(key => {
+      const state = changes[key];
+      const shouldSlam = MongoDbStorage.shouldSlam(state.eTag);
+      const oldETag = state.eTag;
+      state.eTag = new ObjectID().toHexString();
       operations.push({
         updateOne: {
-          filter: {
-            _id: key
-          },
+          filter: MongoDbStorage.createFilter(key, oldETag),
           update: {
             $set: {
-              state: changes[key],
+              state: state,
               dt: new Date()
             }
           },
-          upsert: true
+          upsert: shouldSlam
         }
       })
     })
 
-    await this.Collection.bulkWrite(operations);
-
+    const bulkResults = await this.Collection.bulkWrite(operations);
+    //TODO: process bulk results: if 0 modified, 0 upserted then throw exception because state was not mutated.
   }
 
   public async delete(keys: string[]): Promise<void> {
+    if (!keys || keys.length == 0) {
+      return;
+    }    
     await this.Collection.deleteMany({ _id: { $in: keys } });
+  }
+
+  public static shouldSlam(etag: any) {
+    return (etag === '*' || !etag);
+  }
+
+  public static createFilter(key: string, etag: any) {
+    if (this.shouldSlam(etag)) {
+      return { _id: key };
+    }
+    return { _id: key, 'state.eTag': etag };
   }
 
   get Collection(): Collection<MongoDocumentStoreItem> {
